Big tiler rewrite
Apologies this commit isn't more incremental.

1. Move all of the metadata handling logic out of the tile builder script:
Before DPS jobs are started, we will run a script to generate a plan for the
database (which granules go in which tiles, etc.) based on the region to be
added. The plan will be written as the database metadata.
DPS jobs will only be responsible for reading from database metadata and
loading the requested granules into their assigned tile.
2. Use refreshable credentials for S3 because the DPS jobs are very long-running
3. Add quality filtering
4. Fix lon/lat index slicing -- it is not necessarily the case that
longitudes/latitudes will be strictly increasing/decreasing within a granule.
We have to create a mask of footprints in the tile, rather than slicing.