add a quick spatial filter query helper
While ST_Contains/Intersects/etc. queries can be expensive on a large
database, they are very quick if we can hint to duckdb which files to
look in. (the data is partitioned on tile_id so if we can use this index
it really speeds things up). A geopandas spatial join is well worth it:
- ST_Contains on a ~0.5 degree box over 100 tiles: 1 m 10 s
- WHERE lon_lowestmode < x, ...: 27 s
- gpd.sjoin + AND ({spatial_filter}): 2 s

Note that the gpd.sjoin will not increase in complexity as more files
are added to the database.